Refactor ZigzagDCTcoefficientsOrder to use float instead of int

+++ src/DCT/ZigzagDCTcoefficientsOrder.cpp
@@ -1,8 +1,8 @@
 #include "ZigzagDCTcoefficientsOrder.h"
 
-std::vector<std::vector<int>> ZigzagDCTcoefficientsOrder::ZigZagOrderImageBlocks(std::vector<std::vector<std::vector<int>>> &imageChuncks) const
+std::vector<std::vector<float>> ZigzagDCTcoefficientsOrder::ZigZagOrderImageBlocks(std::vector<std::vector<std::vector<float>>> &imageChuncks) const
 {
-    std::vector<std::vector<int>> results;
+    std::vector<std::vector<float>> results;
 
     for (const auto &ImageBlock : imageChuncks)
     {
@@ -14,10 +14,10 @@ std::vector<std::vector<int>> ZigzagDCTcoefficientsOrder::ZigZagOrderImageBlocks
     return results;
 }
 
-std::vector<int> ZigzagDCTcoefficientsOrder::ZigzagOrder(const std::vector<std::vector<int>> &imageBlock) const
+std::vector<float> ZigzagDCTcoefficientsOrder::ZigzagOrder(const std::vector<std::vector<float>> &imageBlock) const
 {
     int n = static_cast<int>(imageBlock.size());
-    std::vector<int> result;
+    std::vector<float> result;
 
     for (int sum = 0; sum < 2 * n - 1; ++sum)
     {
@@ -27,10 +27,10 @@ std::vector<int> ZigzagDCTcoefficientsOrder::ZigzagOrder(const std::vector<std::
     return result;
 }
 
-std::vector<int> ZigzagDCTcoefficientsOrder::extractZigzagDiagonal(const std::vector<std::vector<int>> &imageBlock, int sum) const
+std::vector<float> ZigzagDCTcoefficientsOrder::extractZigzagDiagonal(const std::vector<std::vector<float>> &imageBlock, int sum) const
 {
     int n = static_cast<int>(imageBlock.size());
-    std::vector<int> diagonal;
+    std::vector<float> diagonal;
     int rowStart = std::max(0, sum - (n - 1));
     int rowEnd = std::min(sum, n - 1);
 
@@ -53,9 +53,9 @@ std::vector<int> ZigzagDCTcoefficientsOrder::extractZigzagDiagonal(const std::ve
     return diagonal;
 }
 
-std::vector<std::vector<std::vector<int>>> ZigzagDCTcoefficientsOrder::deZigZagOrderImageBlocks(std::vector<std::vector<int>> &encodedImageBlocks) const
+std::vector<std::vector<std::vector<float>>> ZigzagDCTcoefficientsOrder::deZigZagOrderImageBlocks(std::vector<std::vector<float>> &encodedImageBlocks) const
 {
-    std::vector<std::vector<std::vector<int>>> decodedImageBlocks;
+    std::vector<std::vector<std::vector<float>>> decodedImageBlocks;
     for (const auto &zigzagOrder : encodedImageBlocks)
     {
         auto imageBlock = ReverseZigzagtraversal(zigzagOrder);
@@ -64,10 +64,10 @@ std::vector<std::vector<std::vector<int>>> ZigzagDCTcoefficientsOrder::deZigZagO
     return decodedImageBlocks;
 }
 
-std::vector<std::vector<int>> ZigzagDCTcoefficientsOrder::ReverseZigzagtraversal(const std::vector<int> &zigzagOrder) const
+std::vector<std::vector<float>> ZigzagDCTcoefficientsOrder::ReverseZigzagtraversal(const std::vector<float> &zigzagOrder) const
 {
     int dctBlockSize = static_cast<int>(std::sqrt(zigzagOrder.size()));
-    std::vector<std::vector<int>> imageBlock(dctBlockSize, std::vector<int>(dctBlockSize, 0));
+    std::vector<std::vector<float>> imageBlock(dctBlockSize, std::vector<float>(dctBlockSize, 0));
     int index = 0;
 
     for (int sum = 0; sum < 2 * dctBlockSize - 1; ++sum)
@@ -78,8 +78,8 @@ std::vector<std::vector<int>> ZigzagDCTcoefficientsOrder::ReverseZigzagtraversal
     return imageBlock;
 }
 
-void ZigzagDCTcoefficientsOrder::fillZigzagDiagonal(std::vector<std::vector<int>> &imageBlock,
-                                                    const std::vector<int> &zigzagOrder,
+void ZigzagDCTcoefficientsOrder::fillZigzagDiagonal(std::vector<std::vector<float>> &imageBlock,
+                                                    const std::vector<float> &zigzagOrder,
                                                     int sum,
                                                     int dctBlockSize,
                                                     int &index) const
