+++ src/DCT/ZigzagDCTcoefficientsOrder.h
@@ -1,7 +1,6 @@
 #ifndef ZIGZAG_DCT_COEFFICIENTS_ORDER_H
 #define ZIGZAG_DCT_COEFFICIENTS_ORDER_H
 
-
 #include "RunLengthEnoding.h"
 #include <vector>
 #include <cmath>
@@ -9,19 +8,19 @@
 class ZigzagDCTcoefficientsOrder
 {
 public:
-    std::vector<std::vector<int>> ZigZagOrderImageBlocks(std::vector<std::vector<std::vector<int>>> &imageChuncks) const;
+    std::vector<std::vector<float>> ZigZagOrderImageBlocks(std::vector<std::vector<std::vector<float>>> &imageChuncks) const;
 
-    std::vector<std::vector<std::vector<int>>> deZigZagOrderImageBlocks(std::vector<std::vector<int>> &imageChuncks) const;
+    std::vector<std::vector<std::vector<float>>> deZigZagOrderImageBlocks(std::vector<std::vector<float>> &imageChuncks) const;
 
 private:
     RunLengthEnoding m_runlengthEnoding;
 
-    void fillZigzagDiagonal(std::vector<std::vector<int>> &imageBlock, const std::vector<int> &zigzagOrder, int sum, int dctBlockSize, int &index) const;
-    std::vector<int> Zigzagtraversal(const std::vector<std::vector<int>> &imageBlock);
-    std::vector<int> ZigzagOrder(const std::vector<std::vector<int>> &imageBlock) const;
-    std::vector<int> extractZigzagDiagonal(const std::vector<std::vector<int>> &imageBlock, int sum) const;
+    void fillZigzagDiagonal(std::vector<std::vector<float>> &imageBlock, const std::vector<float> &zigzagOrder, int sum, int dctBlockSize, int &index) const;
+    std::vector<float> Zigzagtraversal(const std::vector<std::vector<float>> &imageBlock);
+    std::vector<float> ZigzagOrder(const std::vector<std::vector<float>> &imageBlock) const;
+    std::vector<float> extractZigzagDiagonal(const std::vector<std::vector<float>> &imageBlock, int sum) const;
 
-    std::vector<std::vector<int>> ReverseZigzagtraversal(const std::vector<int> &zigzagOrder) const;
+    std::vector<std::vector<float>> ReverseZigzagtraversal(const std::vector<float> &zigzagOrder) const;
 };
 
 #endif // ZIGZAG_DCT_COEFFICIENTS_ORDER_H
